pyproject.toml
[project]
name = "whatsapp-ai-agent"
version = "1.1.0"
description = "WhatsApp AI agent bridge for Evolution API, LLM responses, and Hotmart marketing automation."
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.109.0",
  "uvicorn[standard]==0.27.0",
  "httpx==0.26.0",
  "python-dotenv==1.0.0",
  "fpdf2==2.7.8",
  "gtts==2.5.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "pytest-asyncio==0.24.0",
  "ruff==0.8.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]
